Make sure POSKeyError propagates.
Note that ``queryId`` now propagates this error; previously it was
hidden. This now matches the behaviour of ``queryObject`` which always
propagated.

Test this. I have 100% test coverage locally. Some minor refactoring of
the subscribers facilitates this (and reduces the nesting depth).

Also add a setup.cfg to indicate that we generate universal wheels.