Raise subclasses of KeyError (#5)
* fix testing under py3.3 (could not find 'intid' in 'zope') and simplify

Bring in the .travis.yml changes from #3 and also do the same for tox.ini.

* Raise more informative KeyError subclasses.

* Make sure POSKeyError propagates.

Note that ``queryId`` now propagates this error; previously it was
hidden. This now matches the behaviour of ``queryObject`` which always
propagated.

Test this. I have 100% test coverage locally. Some minor refactoring of
the subscribers facilitates this (and reduces the nesting depth).

Also add a setup.cfg to indicate that we generate universal wheels.

* Use latest pypy3 on travis because of pip.